+++ examples/DummyProcess.ts
@@ -5,15 +5,10 @@ import { DummyFactory } from './DummyFactory';
 
 export class DummyProcess extends ProcessContainer {
 
-    public initReferences(references: IReferences): void {
-        super.initReferences(references);
-
-        // Factory to statically resolve dummy components
-        references.put(DummyFactory.Descriptor, new DummyFactory());
-    }
-
-    public runWithArguments(args: string[]): void {
-        return this.runWithArgumentsOrConfigFile("dummy", args, "./config/dummy.yaml");
+    public constructor() {
+        super("dummy", "Sample dummy process");
+        this._configPath = './config/dummy.yaml';
+        this._factories.add(new DummyFactory());
     }
 
 }