+++ obj/examples/DummyProcess.d.ts
@@ -1,6 +1,4 @@
-import { IReferences } from 'pip-services-commons-node';
 import { ProcessContainer } from '../src/ProcessContainer';
 export declare class DummyProcess extends ProcessContainer {
-    initReferences(references: IReferences): void;
-    runWithArguments(args: string[]): void;
+    constructor();
 }