+++ obj/examples/DummyProcess.js
@@ -3,13 +3,10 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const ProcessContainer_1 = require("../src/ProcessContainer");
 const DummyFactory_1 = require("./DummyFactory");
 class DummyProcess extends ProcessContainer_1.ProcessContainer {
-    initReferences(references) {
-        super.initReferences(references);
-        // Factory to statically resolve dummy components
-        references.put(DummyFactory_1.DummyFactory.Descriptor, new DummyFactory_1.DummyFactory());
-    }
-    runWithArguments(args) {
-        return this.runWithArgumentsOrConfigFile("dummy", args, "./config/dummy.yaml");
+    constructor() {
+        super("dummy", "Sample dummy process");
+        this._configPath = './config/dummy.yaml';
+        this._factories.add(new DummyFactory_1.DummyFactory());
     }
 }
 exports.DummyProcess = DummyProcess;