+++ obj/src/ProcessContainer.d.ts
@@ -1,14 +1,12 @@
-import { ContainerConfig } from './config/ContainerConfig';
 import { Container } from './Container';
 export declare class ProcessContainer extends Container {
-    constructor();
-    private getConfigPath(args, defaultPath);
+    protected _configPath: string;
+    constructor(name?: string, description?: string);
+    private getConfigPath(args);
     private getParameters(args);
     private showHelp(args);
     private printHelp();
     private captureErrors(correlationId);
     private captureExit(correlationId);
-    run(correlationId: string): void;
-    runWithConfig(correlationId: string, config: ContainerConfig): void;
-    runWithArgumentsOrConfigFile(correlationId: string, args: string[], defaultPath: string): void;
+    run(args: string[]): void;
 }